backend/datagen/data_generator_requirements.txt
faker==20.1.0
psycopg2-binary==2.9.9
tabulate==0.9.0
